Move tunnel provider out of wireguard (#1368)
