+++ nym-vpn-core/crates/nym-vpn-lib/src/lib.rs
@@ -10,6 +10,8 @@ mod bandwidth_controller;
 mod error;
 mod mixnet;
 mod platform;
+#[cfg(any(target_os = "ios", target_os = "android"))]
+pub mod tunnel_provider;
 pub mod tunnel_state_machine;
 mod uniffi_custom_impls;
 mod wg_config;
