+++ nym-vpn-core/crates/nym-vpn-lib/src/storage/helpers.rs
@@ -1,3 +1,6 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
 use std::path::{Path, PathBuf};
 
 use nym_vpn_store::keys::{persistence::OnDiskKeysError, DeviceKeys, KeyStore as _};
