+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_provider/ios/default_path_observer.rs
@@ -1,6 +1,10 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
 use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
 
 /// The type alias for the receiving end of path observer.
+#[allow(unused)]
 pub type DefaultPathReceiver = UnboundedReceiver<OSDefaultPath>;
 
 /// Observer type that wraps network path changes into a channel.
