+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/any_tunnel_handle.rs
@@ -58,19 +58,7 @@ impl AnyTunnelHandle {
                     Ok(vec![])
                 }
             },
-            Self::Wireguard(handle) => {
-                #[cfg(any(target_os = "linux", target_os = "macos", target_os = "windows"))]
-                {
-                    let wait_result = handle.wait().await;
-                    Ok(vec![wait_result.entry_tun, wait_result.exit_tun])
-                }
-
-                #[cfg(any(target_os = "ios", target_os = "android"))]
-                {
-                    let tun_device = handle.wait().await;
-                    Ok(vec![tun_device])
-                }
-            }
+            Self::Wireguard(handle) => Ok(handle.wait().await),
         }
     }
 }
