+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/mod.rs
@@ -188,7 +188,7 @@ pub enum Error {
 
     #[cfg(target_os = "ios")]
     #[error("failed to resolve using dns64")]
-    ResolveDns64(#[source] wireguard::ios::dns64::Error),
+    ResolveDns64(#[source] wireguard::dns64::Error),
 
     #[error("failed to open exit connection through the entry tunnel: {0}")]
     OpenExitConnection(#[source] nym_wg_go::Error),
