+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/wireguard/mod.rs
@@ -4,11 +4,7 @@
 pub mod connected_tunnel;
 pub mod connector;
 
-#[cfg(target_os = "android")]
-pub mod android;
 #[cfg(target_os = "ios")]
-pub mod ios;
-#[cfg(any(target_os = "ios", target_os = "android"))]
-pub mod tunnel_settings;
+pub mod dns64;
 #[cfg(any(target_os = "ios", target_os = "android"))]
 pub mod two_hop_config;
