+++ nym-vpn-core/crates/nym-vpnc/src/main.rs
@@ -1,5 +1,5 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: GPL-3.0-only
 
 use anyhow::Result;
 use clap::Parser;
