+++ nym-vpn-core/crates/nym-vpnc/src/protobuf_conversion.rs
@@ -1,5 +1,5 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: GPL-3.0-only
 
 use nym_gateway_directory::{EntryPoint, ExitPoint, GatewayType, NodeIdentity, Recipient};
