+++ nym-vpn-core/crates/nym-wg-go/src/wireguard_go.rs
@@ -1,5 +1,5 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: GPL-3.0-only
 
 #[cfg(unix)]
 use std::os::unix::io::RawFd;
