+++ nym-vpn-core/crates/uniffi-bindgen/src/main.rs
@@ -1,5 +1,5 @@
 // Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: GPL-3.0-only
 
 fn main() {
     uniffi::uniffi_bindgen_main()
